assigns information from input
