+++ app/signupReqFormat/rbuCustomer.js
@@ -1,19 +1,7 @@
-export const CUSTOMER_SIGNUP_REQUEST = {
+export const RBU_SIGNUP_REQUEST = {
   'serviceType': 'SYSOPS',
   'metadata': {
     'property': [
-      {
-        'value': 'true',
-        'key': 'cloudFilesPurchased'
-      },
-      {
-        'value': 'true',
-        'key': 'cloudServersPurchased'
-      },
-      {
-        'value': 'false',
-        'key': 'loadBalancersPurchased'
-      },
       {
         'value': '192.0.2.10',
         'key': 'ipAddress'
@@ -25,11 +13,20 @@ export const CUSTOMER_SIGNUP_REQUEST = {
       {
         'value': '134.288-8901',
         'key': 'deviceFingerPrint'
+      },
+      {
+        'value': 'false',
+        'key': 'skipCloudWelcomeEmail'
+      },
+      {
+        'value': 'RBU',
+        'key': 'Business_Unit'
       }
     ]
   },
-  'accountName': 'soph_racker_123',
+  'accountName': '',
   'description': 'A Karate (RBU) cloud signup request from the retail site.',
+  'skipFraudCheck': true,
   'type': 'CLOUD',
   'termsAndConditions': 'US',
   'acceptTermsAndConditions': true,
@@ -39,8 +36,8 @@ export const CUSTOMER_SIGNUP_REQUEST = {
     'type': 'BRC'
   },
   'promoCode': '788',
-  'vatCode': 'GB0177282',
   'defaultRegion': 'SYD',
+  'externalId': 'AWS',
   'serviceLevel': 'MANAGED',
   'paymentMethod': {
     'paymentInfo': {
@@ -54,8 +51,8 @@ export const CUSTOMER_SIGNUP_REQUEST = {
   'contacts': {
     'contact': [
       {
-        'firstName': '',
-        'lastName': '',
+        'firstName': 'John',
+        'lastName': 'Doe',
         'addresses': {
           'address': [
             {
@@ -71,7 +68,7 @@ export const CUSTOMER_SIGNUP_REQUEST = {
         'emailAddresses': {
           'emailAddress': [
             {
-              'address': 'me@rackspace.com',
+              'address': 'demo@example.com',
               'primary': true
             }
           ]
@@ -82,17 +79,21 @@ export const CUSTOMER_SIGNUP_REQUEST = {
             'PRIMARY'
           ]
         },
-        'title': '',
-        'suffix': '',
+        'suffix': 'Senior',
+        'title': 'Mr',
         'user': {
           'password': 'juv%UBWg@6pS',
-          'username': 'Soph_Racker-123'
+          'secretQA': {
+            'answer': 'Clay',
+            'question': 'What is my name?'
+          },
+          'username': ''
         },
         'phoneNumbers': {
           'phoneNumber': [
             {
               'country': 'US',
-              'number': '1234567890',
+              'number': '6758783848',
               'category': 'HOME',
               'primary': true
             }
