+++ signupform.json
@@ -72,7 +72,7 @@
         },
         "title": "Mr",
         "user": {
-          "password": "Aw3s0meP4ss",
+          "password": "9v3dyXeOLBN$",
           "username": "soph_rbu_user"
         },
         "phoneNumbers": {
